rename  algae manipulator
