+++ components/algae_manipulator.py
